requirements.txt
Flask
openai-whisper
google-generativeai
regex
requests
